requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
python-dotenv
